refactor: 데이터 및 로직을 useRankingList 훅함수로 분리
